+++ gui/merge_tab.py
@@ -1,64 +1,22 @@
 from __future__ import annotations
 
 from typing import List
+import os
 
 from PySide6.QtWidgets import (
-    QWidget, QVBoxLayout, QHBoxLayout, QPushButton, QMessageBox, QLineEdit
+    QWidget, QVBoxLayout, QPushButton, QMessageBox
 )
 
 from utils.i18n import tr, i18n
 from core.drag_drop import DragDropLineEdit
 from core.merge_columns import merge_excel_columns
-
-
-class MappingRow(QWidget):
-    def __init__(self, parent: QWidget | None = None):
-        super().__init__(parent)
-        layout = QHBoxLayout(self)
-        layout.setSpacing(10)
-
-        self.file_input = DragDropLineEdit(mode='file')
-        self.file_input.setPlaceholderText(tr("Файл источник"))
-        layout.addWidget(self.file_input)
-
-        self.src_cols = QLineEdit()
-        self.src_cols.setPlaceholderText(tr("Колонки источника"))
-        layout.addWidget(self.src_cols)
-
-        self.target_sheet = QLineEdit()
-        self.target_sheet.setPlaceholderText(tr("Лист назначения"))
-        layout.addWidget(self.target_sheet)
-
-        self.target_cols = QLineEdit()
-        self.target_cols.setPlaceholderText(tr("Колонки назначения"))
-        layout.addWidget(self.target_cols)
-
-        self.remove_btn = QPushButton("✕")
-        layout.addWidget(self.remove_btn)
-
-    def get_mapping(self):
-        src = self.file_input.text()
-        src_cols = [c.strip() for c in self.src_cols.text().split(',') if c.strip()]
-        tgt_sheet = self.target_sheet.text().strip()
-        tgt_cols = [c.strip() for c in self.target_cols.text().split(',') if c.strip()]
-        return {
-            'source': src,
-            'source_columns': src_cols,
-            'target_sheet': tgt_sheet,
-            'target_columns': tgt_cols,
-        }
-
-    def retranslate(self):
-        self.file_input.setPlaceholderText(tr("Файл источник"))
-        self.src_cols.setPlaceholderText(tr("Колонки источника"))
-        self.target_sheet.setPlaceholderText(tr("Лист назначения"))
-        self.target_cols.setPlaceholderText(tr("Колонки назначения"))
+from .merge_mapping_dialog import MergeMappingDialog
 
 
 class MergeTab(QWidget):
     def __init__(self):
         super().__init__()
-        self.rows: List[MappingRow] = []
+        self.source_files: List[str] = []
         self.init_ui()
         i18n.language_changed.connect(self.retranslate_ui)
         self.retranslate_ui()
@@ -73,16 +31,9 @@ class MergeTab(QWidget):
 
         self.sources_input = DragDropLineEdit(mode='files_or_folder')
         self.sources_input.setPlaceholderText(tr("Файлы источников"))
-        self.sources_input.filesSelected.connect(self.add_rows_from_files)
+        self.sources_input.filesSelected.connect(self.handle_files_selected)
+        self.sources_input.folderSelected.connect(self.handle_folder_selected)
         layout.addWidget(self.sources_input)
-
-        self.mappings_layout = QVBoxLayout()
-        layout.addLayout(self.mappings_layout)
-
-        add_btn = QPushButton(tr("Добавить источник"))
-        add_btn.clicked.connect(self.add_row)
-        layout.addWidget(add_btn)
-        self.add_btn = add_btn
 
         layout.addStretch()
 
@@ -91,47 +42,43 @@ class MergeTab(QWidget):
         layout.addWidget(merge_btn)
         self.merge_btn = merge_btn
 
-    def add_row(self, file_path: str | None = None):
-        row = MappingRow(self)
-        row.remove_btn.clicked.connect(lambda: self.remove_row(row))
-        self.mappings_layout.addWidget(row)
-        self.rows.append(row)
-        if file_path:
-            row.file_input.setText(file_path)
-        return row
+    def handle_files_selected(self, files: List[str]):
+        self.source_files = files
 
-    def remove_row(self, row: MappingRow):
-        row.setParent(None)
-        self.rows.remove(row)
-
-    def add_rows_from_files(self, files: List[str]):
-        for f in files:
-            self.add_row(f)
+    def handle_folder_selected(self, folder: str):
+        excel_exts = ('.xlsx', '.xls')
+        if os.path.isdir(folder):
+            self.source_files = [
+                os.path.join(folder, f)
+                for f in os.listdir(folder)
+                if f.lower().endswith(excel_exts)
+            ]
+        else:
+            self.source_files = []
 
     def run_merge(self):
         main_file = self.main_file_input.text()
         if not main_file:
             QMessageBox.critical(self, tr("Ошибка"), tr("Укажи общий Excel."))
             return
+        if not self.source_files:
+            QMessageBox.critical(self, tr("Ошибка"), tr("Укажи файлы источников."))
+            return
 
-        mappings = []
-        for row in self.rows:
-            mp = row.get_mapping()
-            if not (mp['source'] and mp['source_columns'] and mp['target_sheet'] and mp['target_columns']):
-                QMessageBox.critical(self, tr("Ошибка"), tr("Заполни все поля."))
-                return
-            mappings.append(mp)
-
-        try:
-            output = merge_excel_columns(main_file, mappings)
-            QMessageBox.information(self, tr("Успех"), tr("Файл сохранён: {output}").format(output=output))
-        except Exception as e:
-            QMessageBox.critical(self, tr("Ошибка"), str(e))
+        dialog = MergeMappingDialog(self)
+        for f in self.source_files:
+            dialog.add_row(f)
+        if dialog.exec():
+            mappings = dialog.get_mappings()
+            try:
+                output = merge_excel_columns(main_file, mappings)
+                QMessageBox.information(
+                    self, tr("Успех"), tr("Файл сохранён: {output}").format(output=output)
+                )
+            except Exception as e:
+                QMessageBox.critical(self, tr("Ошибка"), str(e))
 
     def retranslate_ui(self):
         self.main_file_input.setPlaceholderText(tr("Общий Excel"))
         self.sources_input.setPlaceholderText(tr("Файлы источников"))
-        self.add_btn.setText(tr("Добавить источник"))
         self.merge_btn.setText(tr("Объединить"))
-        for r in self.rows:
-            r.retranslate()
